добавил атрибуты для view графика
