+++ core-ui/src/main/res/values/colors.xml
@@ -4,6 +4,7 @@
     <color name="purple_500">#FF6200EE</color>
     <color name="purple_700">#FF3700B3</color>
     <color name="teal_200">#FF03DAC5</color>
+    <color name="teal_200_test">#9003DAC5</color>
     <color name="teal_700">#FF018786</color>
     <color name="black">#000000</color>
     <color name="gray">#9D9D9D</color>
