+++ feature-statistics/src/main/res/values/attr.xml
@@ -1,7 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <declare-styleable name="StatisticsPercentView">
-        <attr name="cornerRadius" format="dimension" />
+        <attr name="corner_radius" format="dimension" />
+    </declare-styleable>
+    <declare-styleable name="WeightChartView">
+        <attr name="wcv_chart_color" format="color" />
+        <attr name="wcv_selected_color" format="color" />
+        <attr name="wcv_skeleton_color" format="color" />
+        <attr name="wcv_text_color" format="color" />
     </declare-styleable>
 </resources>
